Release 1.55: Add documention for the __DATA__ limitation
And some more cross-links.

I'll release that as 1.55 with documented limitation.  And try to actually
fix the __DATA__ limitation in the next release.  I had a similar problem in
B::C and had to treat it seperately, e.g.  here we could run the filter over
the DATA handle and dump it extra.

sort MANIFEST and add MANIFEST.SKIP for .gitignore